+++ src/OpenTelemetry.SemanticConventions/scripts/generate.ps1
@@ -2,8 +2,8 @@ $SCRIPT_DIR = $PSScriptRoot
 $ROOT_DIR = "${SCRIPT_DIR}/../"
 
 # freeze the spec version to make SemanticAttributes generation reproducible
-$SEMCONV_VERSION="1.39.0"
-$GENERATOR_VERSION="v0.20.0"
+$SEMCONV_VERSION="1.41.0"
+$GENERATOR_VERSION="v0.23.0"
 
 Set-Location $SCRIPT_DIR