+++ test/OpenTelemetry.Exporter.Instana.Tests/Processors/ErrorActivityProcessorTests.cs
@@ -10,40 +10,60 @@ namespace OpenTelemetry.Exporter.Instana.Tests.Processors;
 
 public class ErrorActivityProcessorTests
 {
-    private readonly ErrorActivityProcessor errorActivityProcessor = new();
-
     [Fact]
-    public async Task Process_ErrorStatusCodeIsSet()
+    public void Process_ErrorStatusCodeIsSet()
     {
+        // Arrange
         var activity = new Activity("testOperationName");
         activity.SetStatus(ActivityStatusCode.Error, "TestErrorDesc");
         var instanaSpan = new InstanaSpan();
-        await this.errorActivityProcessor.ProcessAsync(activity, instanaSpan);
 
+        // Act
+        var processor = new ErrorActivityProcessor();
+        processor.Process(activity, instanaSpan);
+
+        // Assert
         Assert.Equal(1, instanaSpan.Ec);
         Assert.NotNull(instanaSpan.Data);
-        Assert.NotNull(instanaSpan.Data.data);
-        Assert.Equal("Error", instanaSpan.Data.data[InstanaExporterConstants.ERROR_FIELD]);
-        Assert.Equal("TestErrorDesc", instanaSpan.Data.data[InstanaExporterConstants.ERROR_DETAIL_FIELD]);
+        Assert.NotNull(instanaSpan.Data.Values);
+        Assert.Equal("Error", instanaSpan.Data.Values[InstanaExporterConstants.ErrorField]);
+        Assert.Equal("TestErrorDesc", instanaSpan.Data.Values[InstanaExporterConstants.ErrorDetailField]);
     }
 
     [Fact]
-    public async Task Process_ExistsExceptionEvent()
+    public void Process_ExistsExceptionEvent()
     {
+        // Arrange
         var activity = new Activity("testOperationName");
-        var instanaSpan = new InstanaSpan() { TransformInfo = new Implementation.InstanaSpanTransformInfo() { HasExceptionEvent = true } };
-        await this.errorActivityProcessor.ProcessAsync(activity, instanaSpan);
 
+        var instanaSpan = new InstanaSpan()
+        {
+            TransformInfo = new Implementation.InstanaSpanTransformInfo()
+            {
+                HasExceptionEvent = true,
+            },
+        };
+
+        // Act
+        var processor = new ErrorActivityProcessor();
+        processor.Process(activity, instanaSpan);
+
+        // Assert
         Assert.Equal(1, instanaSpan.Ec);
     }
 
     [Fact]
-    public async Task Process_NoError()
+    public void Process_NoError()
     {
+        // Arrange
         var activity = new Activity("testOperationName");
         var instanaSpan = new InstanaSpan() { TransformInfo = new Implementation.InstanaSpanTransformInfo() };
-        await this.errorActivityProcessor.ProcessAsync(activity, instanaSpan);
 
+        // Act
+        var processor = new ErrorActivityProcessor();
+        processor.Process(activity, instanaSpan);
+
+        // Assert
         Assert.Equal(0, instanaSpan.Ec);
     }
 }